Practice#4: TG bot to send coverage report + Refactor
