+++ .github/workflows/python.yml
@@ -13,18 +13,37 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
+
       - name: Set up Python
         uses: actions/setup-python@v4
         with:
           python-version: '3.12'
+
+      - name: Set up Java
+        uses: actions/setup-java@v3
+        with:
+          distribution: 'adopt'
+          java-version: '11'
+
+      - name: Set up JAVA HOME
+        run: echo "JAVA_HOME=$(dirname $(dirname $(readlink -f $(which java))))" >> $GITHUB_ENV
+
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -r requirements.txt
+          pip install -r requirements.
+
       - name: Run tests
         run: |
+          export JAVA_HOME=$JAVA_HOME
           pytest ./tests --alluredir=allure-results --disable-warnings
         continue-on-error: true
+
+      - name: Send Tests Coverage
+        if: ${{ always() }}
+        run: |
+          python -m packages.notifier.bot
+
       - name: Archive allure results
         uses: actions/upload-artifact@v4
         with:
@@ -42,9 +61,11 @@ jobs:
           distribution: 'microsoft'
           java-version: '17'
       - run: sudo wget https://github.com/allure-framework/allure2/releases/download/2.32.2/allure-2.32.2.tgz && sudo tar -xvzf allure-2.32.2.tgz -C /opt/ && sudo ln -s /opt/allure-2.32.2/bin/allure /usr/bin/allure
+
       - name: Download allure results
         uses: actions/download-artifact@v4
       - run: allure generate -c allure-results -o _site
+
       - name: Store allure report
         uses: actions/upload-artifact@v4
         with:
@@ -59,8 +80,10 @@ jobs:
     steps:
       - name: Download allure results
         uses: actions/download-artifact@v4
+
       - name: Upload allure report
         uses: actions/upload-pages-artifact@v3
+
       - name: Deploy to GitHub Pages
         id: deployment
         uses: actions/deploy-pages@v4
