+++ clients/http/api_mailhog/apis/mailhog_api.py
@@ -1,4 +1,4 @@
-from restclient.client import RestClient
+from packages.restclient.client import RestClient
 
 
 class MailhogApi(RestClient):
