+++ helpers/account_helper.py
@@ -3,11 +3,11 @@ from json import loads
 
 import allure
 
-from dm_api_account.models.change_email import ChangeEmail
-from dm_api_account.models.change_password import ChangePassword
-from dm_api_account.models.login_credentials import LoginCredentials
-from dm_api_account.models.registration import Registration
-from dm_api_account.models.reset_password import ResetPassword
+from clients.http.dm_api_account.models.change_email import ChangeEmail
+from clients.http.dm_api_account.models.change_password import ChangePassword
+from clients.http.dm_api_account.models.login_credentials import LoginCredentials
+from clients.http.dm_api_account.models.registration import Registration
+from clients.http.dm_api_account.models.reset_password import ResetPassword
 from services.api_mailhog import MailHogApi
 from services.dm_api_account import DMApiAccount
