+++ packages/restclient/client.py
@@ -5,12 +5,11 @@ from requests import session
 import structlog
 import uuid
 
-from swagger_coverage_py.listener import CoverageListener
 from swagger_coverage_py.request_schema_handler import RequestSchemaHandler
 from swagger_coverage_py.uri import URI
 
-from restclient.configuration import Configuration
-from restclient.utilities import allure_attach
+from packages.restclient.configuration import Configuration
+from packages.restclient.utilities import allure_attach
 
 
 class RestClient:
