+++ services/api_mailhog.py
@@ -1,5 +1,5 @@
-from api_mailhog.apis.mailhog_api import MailhogApi
-from restclient.configuration import Configuration
+from clients.http.api_mailhog.apis.mailhog_api import MailhogApi
+from packages.restclient.configuration import Configuration
 
 
 class MailHogApi:
