+++ services/dm_api_account.py
@@ -1,6 +1,6 @@
-from dm_api_account.apis.account_api import AccountApi
-from dm_api_account.apis.login_api import LoginApi
-from restclient.configuration import Configuration
+from clients.http.dm_api_account.apis.account_api import AccountApi
+from clients.http.dm_api_account.apis.login_api import LoginApi
+from packages.restclient.configuration import Configuration
 
 
 class DMApiAccount:
