+++ tests/conftest.py
@@ -9,8 +9,9 @@ from swagger_coverage_py.reporter import CoverageReporter
 from vyper import v
 
 from helpers.account_helper import AccountHelper
-from restclient.configuration import Configuration as DmApiConfiguration
-from restclient.configuration import Configuration as MailhogApiConfiguration
+from packages.notifier.bot import send_file
+from packages.restclient.configuration import Configuration as DmApiConfiguration
+from packages.restclient.configuration import Configuration as MailhogApiConfiguration
 from services.api_mailhog import MailHogApi
 from services.dm_api_account import DMApiAccount
 
@@ -35,11 +36,11 @@ options = (
 @pytest.fixture(scope="session", autouse=True)
 def setup_swagger_coverage():
     reporter = CoverageReporter(api_name="dm-api-account", host=v.get("service.dm_api_account"))
-    # reporter.cleanup_input_files()
     reporter.setup("/swagger/Account/swagger.json")
     yield
     reporter.generate_report()
     reporter.cleanup_input_files()
+    send_file()
 
 
 @pytest.fixture(scope="session", autouse=True)
@@ -57,7 +58,6 @@ def set_config(
     os.environ["TELEGRAM_BOT_ACCESS_TOKEN"] = v.get("telegram.token")
     request.config.stash["telegram-notifier-addfields"]["environment"] = config_name
     request.config.stash["telegram-notifier-addfields"]["report"] = "https://dirolorbit.github.io/dm_api_tests/"
-
 
 
 def pytest_addoption(
